Adds a configuration check

+++ BuildScripts/Appveyor/UploadCoveralls.ps1
@@ -1,11 +1,13 @@
-$coverageFilePath = Resolve-Path -path "TestResults\*\*.coverage"
-$coveralls = Resolve-Path -path "packages\*\tools\csmacnz.coveralls.exe.coverage"
+if ($env:CONFIGURATION -eq "Debug"){
+	$coverageFilePath = Resolve-Path -path "TestResults\*\*.coverage"
+	$coveralls = Resolve-Path -path "packages\*\tools\csmacnz.coveralls.exe.coverage"
  
-$coverageFilePath = $coverageFilePath.ToString()
-$coveralls = $coveralls.ToString()
+	$coverageFilePath = $coverageFilePath.ToString()
+	$coveralls = $coveralls.ToString()
  
-if(Test-Path .\coverage.coveragexml){ rm .\coverage.coveragexml }
+	if(Test-Path .\coverage.coveragexml){ rm .\coverage.coveragexml }
  
-& "C:\Program Files (x86)\Microsoft Visual Studio 14.0\Team Tools\Dynamic Code Coverage Tools\CodeCoverage.exe" analyze /output:coverage.coveragexml "$coverageFilePath"
+	& "C:\Program Files (x86)\Microsoft Visual Studio 14.0\Team Tools\Dynamic Code Coverage Tools\CodeCoverage.exe" analyze /output:coverage.coveragexml "$coverageFilePath"
  
-& $coveralls --dynamiccodecoverage -i coverage.coveragexml --useRelativePaths
+	& $coveralls --dynamiccodecoverage -i coverage.coveragexml --useRelativePaths
+}
